LPS-147480 Creation of BaseNotificationsInfo
